Move encoder into the Requests::Support namespace
